+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Change Log
 
 - Initial release
+
+## 2.0.0
+- Fixed the file saving
+- Add Enviroment Variables using .reqenv (this is default), prod.reqenv , dev.reqenv etc. substitute by {{var}} in the editor
+- Fixed multipart file upload